fix for underflow of numMessages

+++ src/libponyrt/actor/actor.h
@@ -40,7 +40,7 @@ typedef struct pony_actor_t
   PONY_ATOMIC(uint8_t) is_muted;
   PONY_ATOMIC(size_t) muted;
   
-  size_t batch;
+  int32_t batch;
 
   // keep things accessed by other actors on a separate cache line
   alignas(64) heap_t heap; // 52/104 bytes
